+++ CHANGELOG.md
@@ -7,6 +7,8 @@ FlatLaf Change Log
   triangle arrows in "Flat IntelliJ" and "Flat Darcula" themes. (issue #7)
 - Added Java 9 module descriptor `module-info.class` to `flatlaf.jar` (in
   `META-INF/versions/9`). But FlatLaf remains Java 8 compatible. (issue #1)
+- Support specifying custom scale factor in system properties `flatlaf.uiScale`
+  or `sun.java2d.uiScale`. E.g. `-Dflatlaf.uiScale=1.5`. (Java 8 only)
 
 
 ## 0.9